boot: Add support for Aruba AP175
This commit adds support for the Aruba AP175.

To bring ethernet up on this board, we have to do some trickery with
the PHY configuration. The IP1001 PHY onboard is able to perform
autonegotiation, and otherwise acts as a fairly generic PHY. However,
it fails to operate if we try to configure it as an AR8033. To get
around this, we lie to U-Boot; in effect, we get it to write register
changes to an absent MDIO1, but read link state from MDIO0 (where the
PHY actually is).

To note, while the main Aruba U-Boot repository[^1] has a generic PHY
driver implementation which apparently covers this PHY[^2] (and even
uses it in the defintion for the MSR2K[^3], this board's 'parent'),
that driver does not actually correctly configure the PHY here.

[^1]: https://github.com/shalzz/aruba-ap-310/blob/master/platform/bootloader/apboot-11n, aka 'apboot'

[^2]: apboot:board/ar7100/common/vsc8601_phy.c

[^3]: apboot:board/ar7100/msr2k/Makefile

Signed-off-by: Martin Kennedy <hurricos@gmail.com>